Remove unused TieredStorageTestHarness
