+++ storage/src/test/java/org/apache/kafka/tiered/storage/README.md
@@ -1,11 +1,11 @@
 # The Test Flow
 
-Step 1: For every test, setup is done via TieredStorageTestHarness which extends IntegrationTestHarness and sets up a cluster with TS enabled on it.
+Step 1: Each test is a standalone class. It declares a `clusterConfig()` method that returns a `ClusterConfig` with tiered storage enabled (via `TieredStorageTestUtils.createServerPropsForRemoteStorage`), and test methods annotated with `@ClusterTemplate("clusterConfig")` that receive a `ClusterInstance` provided by the test framework.
 
-Step 2: The test is written as a specification consisting of sequential actions and assertions. The spec for the complete test is written down first which creates "actions" to be executed.
+Step 2: The test is written as a specification consisting of sequential actions and assertions. The spec for the complete test is built first using `TieredStorageTestBuilder`, which creates the "actions" to be executed.
 
-Step 3: Once we have the test spec in-place (which includes assertion actions), we execute the test which will execute each action sequentially. 
+Step 3: A `TieredStorageTestContext` is created from the `ClusterInstance` (plus any extra consumer config). The test is then executed by running each action of the spec sequentially against the context.
 
-Step 4: The test execution stops when any of the action throws an exception (or an assertion error).
+Step 4: The test execution stops when any of the actions throws an exception (or an assertion error).
 
-Step 5: Clean-up for the test is performed on test exit
+Step 5: Clean-up for the test is performed on test exit — the `TieredStorageTestContext` is closed (it is `AutoCloseable`, typically via try-with-resources) and the test report is printed.
